+++ EssentailFeed/EssentailFeed/Feed Cache/LocalFeedItem.swift
@@ -1,0 +1,22 @@
+//
+//  LocalFeedItem.swift
+//  EssentailFeed
+//
+//  Created by Vishal Wagh on 02/06/22.
+//
+
+import Foundation
+
+public struct LocalFeedItem: Equatable{
+    public let id: UUID
+    public let description: String?
+    public let location: String?
+    public let imageURL: URL
+    
+    public init(id: UUID, description: String?, location: String?, imageURL: URL) {
+        self.id = id
+        self.description = description
+        self.location = location
+        self.imageURL = imageURL
+    }
+}
